Update: Fix issues and assign render by roles

+++ admin/administradores/index.js
@@ -8,6 +8,7 @@ import {
   resetPasswordToUser,
 } from "../../js/config/config.js";
 
+let permissions = {};
 const MODAL_DIALOG = document.querySelector("div.modal-dialog");
 const BUTTON_ADD = document.querySelector("button#button-add");
 const BUTTON_CLOSE_MODAL = document.querySelectorAll(".close-modal");
@@ -62,6 +63,10 @@ const destructureDataFromCollection = (data) => {
 };
 
 const createAdminUser = async () => {
+  if (permissions.create == "0") {
+    negateAccess();
+    return;
+  }
   const formObject = await userReducerFormData();
   const userExists = CONTEXT?.some(
     (item) => item.usuario === formObject.usuario
@@ -201,7 +206,13 @@ async function hideModal(refetch = false) {
   }
 }
 
-BUTTON_ADD.addEventListener("click", async () => showModal("add"));
+BUTTON_ADD.addEventListener("click", async () => {
+  if (permissions.create == "0") {
+    negateAccess();
+    return;
+  }
+  showModal("add");
+});
 
 BUTTON_CLOSE_MODAL.forEach((element) => {
   element.addEventListener("click", async () => hideModal(false));
@@ -250,7 +261,7 @@ function createRowToTable(list) {
 function createActionButtons(uid) {
   const tableCell = document.createElement("td");
   // const viewButton = document.createElement("button");
-  // const resetPasswordButton = document.createElement("button");
+  const resetPasswordButton = document.createElement("button");
   const updateButton = document.createElement("button");
   const deleteButton = document.createElement("button");
 
@@ -262,13 +273,13 @@ function createActionButtons(uid) {
     "data-id": uid,
     title: "Editar",
   };
-  // const resetPasswordAttrs = {
-  //   type: "button",
-  //   class: "cellButtonTool",
-  //   "data-type": "update",
-  //   "data-id": u,
-  //   title: "Reestablecer contraseña",
-  // };
+  const resetPasswordAttrs = {
+    type: "button",
+    class: "cellButtonTool",
+    "data-type": "update",
+    "data-id": uid,
+    title: "Reestablecer contraseña",
+  };
   const deleteAttrs = {
     type: "button",
     class: "cellButtonTool",
@@ -279,33 +290,37 @@ function createActionButtons(uid) {
 
   // elementSetAttributes(viewButton, viewAttrs);
   elementSetAttributes(updateButton, updateAttrs);
-  // elementSetAttributes(resetPasswordButton, resetPasswordAttrs);
+  elementSetAttributes(resetPasswordButton, resetPasswordAttrs);
   elementSetAttributes(deleteButton, deleteAttrs);
   elementSetAttributes(tableCell, { "data-type": "tools" });
 
   // viewButton.innerHTML = "Ver";
   updateButton.innerHTML = `<svg width="16px" height="16px" viewBox="0 0 18 18" xmlns="http://www.w3.org/2000/svg" fill="#000000"><g id="SVGRepo_bgCarrier" stroke-width="0"></g><g id="SVGRepo_tracerCarrier" stroke-linecap="round" stroke-linejoin="round"></g><g id="SVGRepo_iconCarrier"> <path fill="#c88528" d="M2.85 10.907l-.672 1.407L.033 17.26a.535.535 0 0 0 0 .368.917.917 0 0 0 .155.184.917.917 0 0 0 .184.155A.54.54 0 0 0 .56 18a.48.48 0 0 0 .18-.033l4.946-2.145 1.407-.672 8.53-8.53-4.244-4.243zM4.857 14l-1.515.657L4 13.143l.508-1.064 1.415 1.413zM16.707 5.537l-4.244-4.244.707-.707a2 2 0 0 1 2.83 0L17.414 2a2 2 0 0 1 0 2.83z"></path> </g></svg>`;
-  // resetPasswordButton.innerHTML = `<svg width="16px" height="16px" viewBox="0 0 18 18" xmlns="http://www.w3.org/2000/svg" fill="#000000"><g id="SVGRepo_bgCarrier" stroke-width="0"></g><g id="SVGRepo_tracerCarrier" stroke-linecap="round" stroke-linejoin="round"></g><g id="SVGRepo_iconCarrier"> <path d="M17.408 3.412a1.974 1.974 0 0 0 0-2.82 1.973 1.973 0 0 0-2.819 0l-.29.29-.59-.59a1.009 1.009 0 0 0-1.65.35l-.35-.35a1.004 1.004 0 1 0-1.42 1.42l.35.35a1.033 1.033 0 0 0-.58.58l-.35-.35a1.004 1.004 0 0 0-1.42 1.42L9.879 5.3l-3.02 3.01c-.01.01-.02.03-.03.04A4.885 4.885 0 0 0 5 8a5 5 0 1 0 5 5 4.885 4.885 0 0 0-.35-1.83c.01-.01.03-.02.04-.03l7.718-7.728zM5 15a2 2 0 1 1 0-4 2 2 0 0 1 0 4z" fill="#287bc8" fill-rule="evenodd"></path> </g></svg>`;
+  resetPasswordButton.innerHTML = `<svg width="16px" height="16px" viewBox="0 0 18 18" xmlns="http://www.w3.org/2000/svg" fill="#000000"><g id="SVGRepo_bgCarrier" stroke-width="0"></g><g id="SVGRepo_tracerCarrier" stroke-linecap="round" stroke-linejoin="round"></g><g id="SVGRepo_iconCarrier"> <path d="M17.408 3.412a1.974 1.974 0 0 0 0-2.82 1.973 1.973 0 0 0-2.819 0l-.29.29-.59-.59a1.009 1.009 0 0 0-1.65.35l-.35-.35a1.004 1.004 0 1 0-1.42 1.42l.35.35a1.033 1.033 0 0 0-.58.58l-.35-.35a1.004 1.004 0 0 0-1.42 1.42L9.879 5.3l-3.02 3.01c-.01.01-.02.03-.03.04A4.885 4.885 0 0 0 5 8a5 5 0 1 0 5 5 4.885 4.885 0 0 0-.35-1.83c.01-.01.03-.02.04-.03l7.718-7.728zM5 15a2 2 0 1 1 0-4 2 2 0 0 1 0 4z" fill="#287bc8" fill-rule="evenodd"></path> </g></svg>`;
   deleteButton.innerHTML = `<svg width="16px" height="16px" viewBox="0 0 18 18" xmlns="http://www.w3.org/2000/svg" fill="#000000"><g id="SVGRepo_bgCarrier" stroke-width="0"></g><g id="SVGRepo_tracerCarrier" stroke-linecap="round" stroke-linejoin="round"></g><g id="SVGRepo_iconCarrier"> <path fill="#c84028" d="M13 18H5a2 2 0 0 1-2-2V7a1 1 0 0 1 1-1h10a1 1 0 0 1 1 1v9a2 2 0 0 1-2 2zm3-15a1 1 0 0 1-1 1H3a1 1 0 0 1 0-2h3V1a1 1 0 0 1 1-1h4a1 1 0 0 1 1 1v1h3a1 1 0 0 1 1 1z"></path> </g></svg>`;
 
   updateButton.addEventListener(
     "click",
     async (e) => await handleUpdateUser(uid)
   );
-  // resetPasswordButton.addEventListener("click", (e) =>
-  //   handleResetPassword(e.currentTarget.parentNode)
-  // );
+  resetPasswordButton.addEventListener("click", (e) =>
+    handleResetPassword(uid)
+  );
   deleteButton.addEventListener("click", (e) => handleDeleteUser(uid));
 
   // tableCell.insertAdjacentElement("beforeend", viewButton);
   tableCell.insertAdjacentElement("beforeend", updateButton);
-  // tableCell.insertAdjacentElement("beforeend", resetPasswordButton);
+  tableCell.insertAdjacentElement("beforeend", resetPasswordButton);
   tableCell.insertAdjacentElement("beforeend", deleteButton);
 
   return tableCell;
 }
 
 async function handleUpdateUser(uid) {
+  if (permissions.update == "0") {
+    negateAccess();
+    return;
+  }
   showModal("update");
   MODAL_DIALOG.setAttribute("data-document", uid);
   let data = await getDocumentByUid(uid, "administradores");
@@ -321,6 +336,10 @@ async function handleUpdateUser(uid) {
 }
 
 async function handleDeleteUser(uid) {
+  if (permissions.delete == "0") {
+    negateAccess();
+    return;
+  }
   const message = "Estas seguro de eliminar a este usuario?";
   if (confirm(message)) {
     const response = await deleteDocumentByUid(uid, "administradores");
@@ -337,6 +356,9 @@ function elementSetAttributes(el, attrs) {
 }
 
 const getRolesList = async () => {
+  if (permissions.update == "0" && permissions.create == "0") {
+    return;
+  }
   try {
     const selectElement = document.querySelector("select#rolClave");
     const roles = await getDocumentsFromCollection("roles");
@@ -349,7 +371,7 @@ const getRolesList = async () => {
         "</option>";
       selectElement.insertAdjacentHTML("beforeend", optionElement);
       rolesList.push({ uid: document.id, ...document.data() });
-    }); 
+    });
   } catch (error) {
     console.error("No se pudo obtener datos de roles", error);
   }
@@ -403,5 +425,28 @@ const errorHandler = (code, args) => {
 //   document.querySelector("#usuario").value = username;
 // }
 
-getRolesList();
-fetchData();
+const rolesPermission = async () => {
+  const permission = sessionStorage.getItem("adminsPermission");
+  if (permission === null || permission === undefined) {
+    window.location.replace("../index.html");
+  }
+
+  permission.split("$").forEach((permit) => {
+    const name = permit.split(",")[0];
+    const value = permit.split(",")[1];
+    console.log({ name, value });
+    permissions[name] = value;
+  });
+};
+
+const negateAccess = () => {
+  window.alert(
+    "No cuentas con los permisos necesarios para realizar esta operación."
+  );
+};
+
+document.addEventListener("DOMContentLoaded", async function () {
+  await rolesPermission();
+  await getRolesList();
+  await fetchData();
+});
